consumer/requirements.txt
mongoengine
flask_login
flask
pyjwt
pillow
gunicorn==20.0.4
